Refactor Card Table into a separate component
